Add LocaleTrait to reduce code duplication
