+++ src/Geocoder/Provider/Yandex.php
@@ -12,6 +12,7 @@ namespace Geocoder\Provider;
 
 use Geocoder\Exception\NoResult;
 use Geocoder\Exception\UnsupportedOperation;
+use Geocoder\Provider\LocaleTrait;
 use Ivory\HttpAdapter\HttpAdapterInterface;
 
 /**
@@ -19,6 +20,8 @@ use Ivory\HttpAdapter\HttpAdapterInterface;
  */
 class Yandex extends AbstractHttpProvider implements LocaleAwareProvider
 {
+    use LocaleTrait;
+
     /**
      * @var string
      */
@@ -33,11 +36,6 @@ class Yandex extends AbstractHttpProvider implements LocaleAwareProvider
      * @var string
      */
     private $toponym;
-
-    /**
-     * @var string
-     */
-    private $locale;
 
     /**
      * @param HttpAdapterInterface $adapter An HTTP adapter.
@@ -87,24 +85,6 @@ class Yandex extends AbstractHttpProvider implements LocaleAwareProvider
     public function getName()
     {
         return 'yandex';
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getLocale()
-    {
-        return $this->locale;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function setLocale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
     }
 
     /**
